+++ planning/collision_free_path_planner/include/collision_free_path_planner/utils/cv_utils.hpp
@@ -29,6 +29,8 @@
 #include <memory>
 #include <vector>
 
+namespace collision_free_path_planner
+{
 namespace util
 {
 struct Footprint;
@@ -117,5 +119,5 @@ bool isOutsideDrivableAreaFromCirclesFootprint(
   const std::vector<double> vehicle_circle_longitudinal_offsets,
   const double vehicle_circle_radius);
 }  // namespace cv_drivable_area_utils
-
+}  // namespace collision_free_path_planner
 #endif  // COLLISION_FREE_PATH_PLANNER__UTILS__CV_UTILS_HPP_
